add time features

+++ Sources/AsyncConnection.swift
@@ -1,3 +1,3 @@
 public protocol AsyncConnection: AsyncStream {
-    func open(timingOut deadline: Int64, result: (Void throws -> AsyncConnection) -> Void) throws
+    func open(timingOut deadline: Double, result: (Void throws -> AsyncConnection) -> Void) throws
 }
