+++ Sources/AsyncHost.swift
@@ -1,3 +1,3 @@
 public protocol AsyncHost {
-    func accept(timingOut deadline: Int64, result: (Void throws -> AsyncStream) -> Void)
+    func accept(timingOut deadline: Double, result: (Void throws -> AsyncStream) -> Void)
 }
