+++ Sources/Connection.swift
@@ -1,3 +1,3 @@
 public protocol Connection: Stream {
-    func open(timingOut deadline: Int64) throws
+    func open(timingOut deadline: Double) throws
 }
