+++ Sources/Stream.swift
@@ -1,10 +1,10 @@
 public protocol Sending: AsyncSending {
-    func send(data: Data, timingOut deadline: Int64) throws
-    func flush(timingOut deadline: Int64) throws
+    func send(data: Data, timingOut deadline: Double) throws
+    func flush(timingOut deadline: Double) throws
 }
 
 public protocol Receiving: AsyncReceiving {
-    func receive(upTo byteCount: Int, timingOut deadline: Int64) throws -> Data
+    func receive(upTo byteCount: Int, timingOut deadline: Double) throws -> Data
 }
 
 public protocol SendingStream: Closable, Sending {}
@@ -12,17 +12,17 @@ public protocol ReceivingStream: Closable, Receiving {}
 public protocol Stream: SendingStream, ReceivingStream {}
 
 extension Sending {
-    public func send(data: Data, timingOut deadline: Int64, result: (Void throws -> Void) -> Void) {
+    public func send(data: Data, timingOut deadline: Double, result: (Void throws -> Void) -> Void) {
         result { try self.send(data, timingOut: deadline) }
     }
 
-    public func flush(timingOut deadline: Int64, result: (Void throws -> Void) -> Void) {
+    public func flush(timingOut deadline: Double, result: (Void throws -> Void) -> Void) {
         result { try self.flush(timingOut: deadline) }
     }
 }
 
 extension Receiving {
-    public func receive(upTo byteCount: Int, timingOut deadline: Int64, result: (Void throws -> Data) -> Void) {
+    public func receive(upTo byteCount: Int, timingOut deadline: Double, result: (Void throws -> Data) -> Void) {
         result { try self.receive(upTo: byteCount, timingOut: deadline) }
     }
 }
